requirements/dev.txt
-r base.txt
django-debug-toolbar==1.3.0
django-extensions==1.5.2
pip-tools==0.3.6
six==1.9.0
sqlparse==0.1.15
Werkzeug==0.10.4
